pyproject.toml
[tool.poetry]
name = "quonfig-openfeature"
version = "0.1.0"
description = "OpenFeature provider for Quonfig — Python server-side SDK"
license = "MIT"
authors = ["Quonfig <hello@quonfig.com>"]
readme = "README.md"
packages = [{include = "quonfig_openfeature"}]

[tool.poetry.dependencies]
python = "^3.9"
# Runtime dependency on the Quonfig Python SDK. For local development we
# resolve against the sibling repo via the `dev` group below; published
# packages will resolve against the public `quonfig` distribution.
quonfig = ">=0.0.7"
openfeature-sdk = ">=0.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-mock = "^3.0"
ruff = "^0.4"
mypy = "^1.0"
# Path dependency so the sibling sdk-python is what actually gets used in
# development and CI. `develop = true` keeps it as an editable install.
quonfig = {path = "../sdk-python", develop = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
